Add Payload CMS dependencies and related packages
