+++ .env.example
@@ -1,0 +1,12 @@
+# Payload CMS Configuration
+PAYLOAD_SECRET=your-secret-key-here
+MONGODB_URI=mongodb://localhost:27017/battleground
+
+# AWS Configuration (existing)
+# AWS_ACCESS_KEY_ID=your-aws-access-key
+# AWS_SECRET_ACCESS_KEY=your-aws-secret-key
+# AWS_REGION=us-east-1
+
+# Other configuration
+# NEXT_PUBLIC_CLERK_PUBLISHABLE_KEY=
+# CLERK_SECRET_KEY=
