+++ payload.config.ts
@@ -1,0 +1,28 @@
+import { mongooseAdapter } from '@payloadcms/db-mongodb'
+import { lexicalEditor } from '@payloadcms/richtext-lexical'
+import { buildConfig } from 'payload'
+import { TextModels } from './src/payload/collections/TextModels'
+
+export default buildConfig({
+  admin: {
+    user: 'users',
+  },
+  collections: [
+    {
+      slug: 'users',
+      auth: true,
+      fields: [
+        // Add custom user fields here if needed
+      ],
+    },
+    TextModels,
+  ],
+  editor: lexicalEditor({}),
+  secret: process.env.PAYLOAD_SECRET || 'your-secret-here',
+  typescript: {
+    outputFile: 'src/payload/types.ts',
+  },
+  db: mongooseAdapter({
+    url: process.env.MONGODB_URI || 'mongodb://localhost:27017/battleground',
+  }),
+})
