+++ src/app/api/[[...slug]]/route.ts
@@ -1,0 +1,8 @@
+/* THIS FILE WAS GENERATED AUTOMATICALLY BY PAYLOAD. */
+import configPromise from '@payload-config'
+import { REST_DELETE, REST_GET, REST_PATCH, REST_POST } from '@payloadcms/next/rest'
+
+export const GET = REST_GET(configPromise)
+export const POST = REST_POST(configPromise)
+export const DELETE = REST_DELETE(configPromise)
+export const PATCH = REST_PATCH(configPromise)
