+++ src/lib/model/models-static.ts
@@ -1,0 +1,591 @@
+import {
+  ai21JambaInstructConfig,
+  anthropicClaudeDefaultConfig,
+  anthropicClaudeSonnet35Config,
+  antrhopicClaudeSonnet37Config,
+  cohereCommandRModelConfig,
+  deepseekR1Config,
+  llamaDefaultModelConfig,
+  lumaRay2ModelConfig,
+  mistral7bInstructConfig,
+  mistral8x7bInstructConfig,
+  mistralLargeConfig,
+  novaConfig,
+  o3MiniConfig,
+  openAiGpt4oConfig,
+  openAiGpt4ominiConfig,
+  stabilityModelConfig,
+  titanImageModelConfig,
+  titanTextExpressConfig,
+  titanTextLiteConfig,
+  titanTextPremierConfig,
+  writerModelConfig,
+} from "./model-configs";
+import { ImageModel, TextModel, VideoModel } from "./model.type";
+
+export const textModels: TextModel[] = [
+  {
+    provider: "Amazon",
+    id: "amazon.titan-text-lite-v1",
+    name: "Titan Text Lite",
+    inputCostPerToken: 0.00015 / 1e3,
+    outputCostPerToken: 0.0002 / 1e3,
+    inputModalities: ["TEXT"],
+    outputModalities: ["TEXT"],
+    config: titanTextLiteConfig,
+    systemPromptSupport: false,
+  },
+  {
+    provider: "Amazon",
+    id: "amazon.titan-text-express-v1",
+    name: "Titan Text Express",
+    inputCostPerToken: 0.0002 / 1e3,
+    outputCostPerToken: 0.0006 / 1e3,
+    inputModalities: ["TEXT"],
+    outputModalities: ["TEXT"],
+    config: titanTextExpressConfig,
+    systemPromptSupport: false,
+  },
+  {
+    provider: "Amazon",
+    id: "amazon.titan-text-premier-v1:0",
+    region: "us-east-1",
+    name: "Titan Text Premier",
+    inputCostPerToken: 0.0005 / 1e3,
+    outputCostPerToken: 0.0015 / 1e3,
+    inputModalities: ["TEXT"],
+    outputModalities: ["TEXT"],
+    config: titanTextPremierConfig,
+    systemPromptSupport: false,
+  },
+  {
+    provider: "Amazon",
+    id: "us.amazon.nova-micro-v1:0",
+    region: "us-east-1",
+    name: "Nova Micro",
+    inputCostPerToken: 0.0000035 / 1e3,
+    outputCostPerToken: 0.00014 / 1e3,
+    inputModalities: ["TEXT"],
+    outputModalities: ["TEXT"],
+    config: novaConfig,
+    systemPromptSupport: true,
+  },
+  {
+    provider: "Amazon",
+    id: "us.amazon.nova-lite-v1:0",
+    region: "us-east-1",
+    name: "Nova Lite",
+    inputCostPerToken: 0.00006 / 1e3,
+    outputCostPerToken: 0.00024 / 1e3,
+    inputModalities: ["TEXT", "IMAGE"],
+    outputModalities: ["TEXT"],
+    config: novaConfig,
+    systemPromptSupport: true,
+  },
+  {
+    provider: "Amazon",
+    id: "us.amazon.nova-pro-v1:0",
+    region: "us-east-1",
+    name: "Nova Pro",
+    inputCostPerToken: 0.0008 / 1e3,
+    outputCostPerToken: 0.0032 / 1e3,
+    inputModalities: ["TEXT", "IMAGE"],
+    outputModalities: ["TEXT"],
+    config: novaConfig,
+    systemPromptSupport: true,
+  },
+  {
+    provider: "Amazon",
+    id: "us.amazon.nova-premier-v1:0",
+    region: "us-east-1",
+    name: "Nova Premier",
+    inputCostPerToken: 0.0025 / 1e3,
+    outputCostPerToken: 0.0125 / 1e3,
+    inputModalities: ["TEXT", "IMAGE"],
+    outputModalities: ["TEXT"],
+    config: novaConfig,
+    systemPromptSupport: true,
+  },
+  {
+    provider: "Anthropic",
+    id: "anthropic.claude-3-haiku-20240307-v1:0",
+    name: "Claude 3 Haiku",
+    inputCostPerToken: 0.00025 / 1e3,
+    outputCostPerToken: 0.00125 / 1e3,
+    inputModalities: ["TEXT"],
+    outputModalities: ["TEXT"],
+    config: anthropicClaudeDefaultConfig,
+    systemPromptSupport: true,
+  },
+  {
+    provider: "Anthropic",
+    id: "anthropic.claude-3-5-haiku-20241022-v1:0",
+    name: "Claude 3.5 Haiku",
+    region: "us-west-2",
+    inputCostPerToken: 0.0008 / 1e3,
+    outputCostPerToken: 0.0004 / 1e3,
+    inputModalities: ["TEXT"],
+    outputModalities: ["TEXT"],
+    config: anthropicClaudeDefaultConfig,
+    systemPromptSupport: true,
+  },
+  {
+    provider: "Anthropic",
+    id: "anthropic.claude-3-sonnet-20240229-v1:0",
+    name: "Claude 3 Sonnet",
+    inputCostPerToken: 0.003 / 1e3,
+    outputCostPerToken: 0.015 / 1e3,
+    inputModalities: ["TEXT", "IMAGE"],
+    outputModalities: ["TEXT"],
+    config: anthropicClaudeDefaultConfig,
+    systemPromptSupport: true,
+  },
+  {
+    provider: "Anthropic",
+    id: "anthropic.claude-3-opus-20240229-v1:0",
+    name: "Claude 3 Opus",
+    region: "us-west-2",
+    inputCostPerToken: 0.015 / 1e3,
+    outputCostPerToken: 0.075 / 1e3,
+    inputModalities: ["TEXT", "IMAGE"],
+    outputModalities: ["TEXT"],
+    config: anthropicClaudeDefaultConfig,
+    systemPromptSupport: true,
+  },
+  {
+    provider: "Anthropic",
+    id: "anthropic.claude-3-5-sonnet-20240620-v1:0",
+    name: "Claude 3.5 Sonnet",
+    inputCostPerToken: 0.003 / 1e3,
+    outputCostPerToken: 0.015 / 1e3,
+    inputModalities: ["TEXT", "IMAGE"],
+    outputModalities: ["TEXT"],
+    config: anthropicClaudeSonnet35Config,
+    systemPromptSupport: true,
+  },
+  {
+    provider: "Anthropic",
+    id: "anthropic.claude-3-5-sonnet-20241022-v2:0",
+    region: "us-west-2",
+    name: "Claude 3.5 Sonnet V2",
+    inputCostPerToken: 0.003 / 1e3,
+    outputCostPerToken: 0.015 / 1e3,
+    inputModalities: ["TEXT", "IMAGE"],
+    outputModalities: ["TEXT"],
+    config: anthropicClaudeSonnet35Config,
+    systemPromptSupport: true,
+  },
+  {
+    provider: "Anthropic",
+    id: "us.anthropic.claude-3-7-sonnet-20250219-v1:0",
+    name: "Claude 3.7 Sonnet",
+    inputCostPerToken: 0.003 / 1e3,
+    outputCostPerToken: 0.015 / 1e3,
+    inputModalities: ["TEXT", "IMAGE"],
+    outputModalities: ["TEXT"],
+    config: antrhopicClaudeSonnet37Config,
+    systemPromptSupport: true,
+  },
+  {
+    provider: "Anthropic",
+    id: "us.anthropic.claude-sonnet-4-20250514-v1:0",
+    region: "us-east-1",
+    name: "Claude 4 Sonnet",
+    inputCostPerToken: 0.003 / 1e3,
+    outputCostPerToken: 0.015 / 1e3,
+    inputModalities: ["TEXT", "IMAGE"],
+    outputModalities: ["TEXT"],
+    config: anthropicClaudeSonnet35Config,
+    systemPromptSupport: true,
+  },
+  {
+    provider: "Anthropic",
+    id: "us.anthropic.claude-opus-4-20250514-v1:0",
+    region: "us-east-1",
+    name: "Claude 4 Opus",
+    inputCostPerToken: 0.015 / 1e3,
+    outputCostPerToken: 0.075 / 1e3,
+    inputModalities: ["TEXT", "IMAGE"],
+    outputModalities: ["TEXT"],
+    config: anthropicClaudeSonnet35Config,
+    systemPromptSupport: true,
+  },
+  {
+    provider: "Deepseek",
+    id: "us.deepseek.r1-v1:0",
+    name: "Deepseek R1",
+    region: "us-west-2",
+    inputCostPerToken: 0.00135 / 1e3,
+    outputCostPerToken: 0.0054 / 1e3,
+    inputModalities: ["TEXT"],
+    outputModalities: ["TEXT"],
+    config: deepseekR1Config,
+    systemPromptSupport: true,
+  },
+  {
+    provider: "Cohere",
+    id: "cohere.command-r-v1:0",
+    name: "Command R",
+    region: "us-west-2",
+    inputCostPerToken: 0.0005 / 1e3,
+    outputCostPerToken: 0.0015 / 1e3,
+    inputModalities: ["TEXT"],
+    outputModalities: ["TEXT"],
+    config: cohereCommandRModelConfig,
+    systemPromptSupport: true,
+  },
+  {
+    provider: "Cohere",
+    id: "cohere.command-r-plus-v1:0",
+    name: "Command R+",
+    inputCostPerToken: 0.003 / 1e3,
+    outputCostPerToken: 0.015 / 1e3,
+    inputModalities: ["TEXT"],
+    outputModalities: ["TEXT"],
+    config: cohereCommandRModelConfig,
+    systemPromptSupport: true,
+  },
+  {
+    provider: "Meta",
+    id: "meta.llama3-8b-instruct-v1:0",
+    name: "Llama 3 8B Instruct",
+    inputCostPerToken: 0.0003 / 1e3,
+    outputCostPerToken: 0.0006 / 1e3,
+    inputModalities: ["TEXT"],
+    outputModalities: ["TEXT"],
+    config: llamaDefaultModelConfig,
+    systemPromptSupport: true,
+  },
+  {
+    provider: "Meta",
+    id: "meta.llama3-70b-instruct-v1:0",
+    name: "Llama 3 70B Instruct",
+    inputCostPerToken: 0.00265 / 1e3,
+    outputCostPerToken: 0.0035 / 1e3,
+    inputModalities: ["TEXT"],
+    outputModalities: ["TEXT"],
+    config: llamaDefaultModelConfig,
+    systemPromptSupport: true,
+  },
+  {
+    provider: "Meta",
+    id: "meta.llama3-1-8b-instruct-v1:0",
+    name: "Llama 3.1 8B Instruct",
+    region: "us-west-2",
+    inputCostPerToken: 0.00022 / 1e3,
+    outputCostPerToken: 0.00022 / 1e3,
+    inputModalities: ["TEXT"],
+    outputModalities: ["TEXT"],
+    config: llamaDefaultModelConfig,
+    systemPromptSupport: true,
+  },
+  {
+    provider: "Meta",
+    id: "meta.llama3-1-70b-instruct-v1:0",
+    name: "Llama 3.1 70B Instruct",
+    region: "us-west-2",
+    inputCostPerToken: 0.00072 / 1e3,
+    outputCostPerToken: 0.00072 / 1e3,
+    inputModalities: ["TEXT"],
+    outputModalities: ["TEXT"],
+    config: llamaDefaultModelConfig,
+    systemPromptSupport: true,
+  },
+  {
+    provider: "Meta",
+    id: "meta.llama3-1-405b-instruct-v1:0",
+    name: "Llama 3.1 405B Instruct",
+    region: "us-west-2",
+    inputCostPerToken: 0.0024 / 1e3,
+    outputCostPerToken: 0.0024 / 1e3,
+    inputModalities: ["TEXT"],
+    outputModalities: ["TEXT"],
+    config: llamaDefaultModelConfig,
+    systemPromptSupport: true,
+  },
+  {
+    provider: "Meta",
+    id: "us.meta.llama3-2-1b-instruct-v1:0",
+    name: "Llama 3.2 1B Instruct",
+    region: "us-west-2",
+    inputCostPerToken: 0.0001 / 1e3,
+    outputCostPerToken: 0.0001 / 1e3,
+    inputModalities: ["TEXT"],
+    outputModalities: ["TEXT"],
+    config: llamaDefaultModelConfig,
+    systemPromptSupport: true,
+  },
+  {
+    provider: "Meta",
+    id: "us.meta.llama3-2-3b-instruct-v1:0",
+    name: "Llama 3.2 3B Instruct",
+    region: "us-west-2",
+    inputCostPerToken: 0.00015 / 1e3,
+    outputCostPerToken: 0.00015 / 1e3,
+    inputModalities: ["TEXT"],
+    outputModalities: ["TEXT"],
+    config: llamaDefaultModelConfig,
+    systemPromptSupport: true,
+  },
+  {
+    provider: "Meta",
+    id: "us.meta.llama3-2-11b-instruct-v1:0",
+    name: "Llama 3.2 11B Instruct",
+    region: "us-west-2",
+    inputCostPerToken: 0.00016 / 1e3,
+    outputCostPerToken: 0.00016 / 1e3,
+    inputModalities: ["TEXT", "IMAGE"],
+    outputModalities: ["TEXT"],
+    config: llamaDefaultModelConfig,
+    systemPromptSupport: true,
+  },
+  {
+    provider: "Meta",
+    id: "us.meta.llama3-2-90b-instruct-v1:0",
+    name: "Llama 3.2 90B Instruct",
+    region: "us-west-2",
+    inputCostPerToken: 0.00072 / 1e3,
+    outputCostPerToken: 0.00072 / 1e3,
+    inputModalities: ["TEXT", "IMAGE"],
+    outputModalities: ["TEXT"],
+    config: llamaDefaultModelConfig,
+    systemPromptSupport: true,
+  },
+  {
+    provider: "Meta",
+    id: "us.meta.llama3-3-70b-instruct-v1:0",
+    name: "Llama 3.3 70B Instruct",
+    region: "us-west-2",
+    inputCostPerToken: 0.00072 / 1e3,
+    outputCostPerToken: 0.00072 / 1e3,
+    inputModalities: ["TEXT"],
+    outputModalities: ["TEXT"],
+    config: llamaDefaultModelConfig,
+    systemPromptSupport: true,
+  },
+  {
+    provider: "Meta",
+    id: "us.meta.llama4-maverick-17b-instruct-v1:0",
+    name: "Llama 4 Maverick 17B Instruct",
+    inputCostPerToken: 0.00024 / 1e3,
+    outputCostPerToken: 0.00097 / 1e3,
+    inputModalities: ["TEXT", "IMAGE"],
+    outputModalities: ["TEXT"],
+    config: llamaDefaultModelConfig,
+    systemPromptSupport: true,
+  },
+  {
+    provider: "Meta",
+    id: "us.meta.llama4-scout-17b-instruct-v1:0",
+    name: "Llama 4 Scout 17B Instruct",
+    inputCostPerToken: 0.00017 / 1e3,
+    outputCostPerToken: 0.00066 / 1e3,
+    inputModalities: ["TEXT", "IMAGE"],
+    outputModalities: ["TEXT"],
+    config: llamaDefaultModelConfig,
+    systemPromptSupport: true,
+  },
+  {
+    provider: "Mistral",
+    id: "mistral.mistral-7b-instruct-v0:2",
+    name: "Mistral 7B Instruct",
+    inputCostPerToken: 0.00015 / 1e3,
+    outputCostPerToken: 0.0002 / 1e3,
+    inputModalities: ["TEXT"],
+    outputModalities: ["TEXT"],
+    config: mistral7bInstructConfig,
+    systemPromptSupport: false,
+  },
+  {
+    provider: "Mistral",
+    id: "mistral.mixtral-8x7b-instruct-v0:1",
+    name: "Mixtral 8x7B Instruct",
+    inputCostPerToken: 0.00045 / 1e3,
+    outputCostPerToken: 0.0007 / 1e3,
+    inputModalities: ["TEXT"],
+    outputModalities: ["TEXT"],
+    config: mistral8x7bInstructConfig,
+    systemPromptSupport: false,
+  },
+  {
+    provider: "Mistral",
+    id: "mistral.mistral-large-2407-v1:0",
+    name: "Mistral Large v2",
+    region: "us-west-2",
+    inputCostPerToken: 0.002 / 1e3,
+    outputCostPerToken: 0.006 / 1e3,
+    inputModalities: ["TEXT"],
+    outputModalities: ["TEXT"],
+    config: mistralLargeConfig,
+    systemPromptSupport: true,
+  },
+  {
+    provider: "Nvidia",
+    id: "meta/llama3-8b-instruct",
+    name: "Llama 3 8B Instruct (Nvidia NIM)",
+    inputModalities: ["TEXT"],
+    outputModalities: ["TEXT"],
+    config: llamaDefaultModelConfig,
+    systemPromptSupport: true,
+  },
+  {
+    provider: "Nvidia",
+    id: "meta/llama3-70b-instruct",
+    name: "Llama 3 70B Instruct (Nvidia NIM)",
+    inputModalities: ["TEXT"],
+    outputModalities: ["TEXT"],
+    config: llamaDefaultModelConfig,
+    systemPromptSupport: true,
+  },
+  {
+    provider: "OpenAI",
+    id: "gpt-4o",
+    name: "GPT 4o",
+    inputCostPerToken: 0.0025 / 1e3,
+    outputCostPerToken: 0.01 / 1e3,
+    inputModalities: ["TEXT", "IMAGE"],
+    outputModalities: ["TEXT"],
+    config: openAiGpt4oConfig,
+    systemPromptSupport: true,
+  },
+  {
+    provider: "OpenAI",
+    id: "gpt-4o-mini",
+    name: "GPT 4o Mini",
+    inputCostPerToken: 0.00015 / 1e3,
+    outputCostPerToken: 0.0006 / 1e3,
+    inputModalities: ["TEXT", "IMAGE"],
+    outputModalities: ["TEXT"],
+    config: openAiGpt4ominiConfig,
+    systemPromptSupport: true,
+  },
+  {
+    provider: "OpenAI",
+    id: "o3-mini",
+    name: "o3 Mini",
+    inputCostPerToken: 0.0011 / 1e3,
+    outputCostPerToken: 0.0044 / 1e3,
+    inputModalities: ["TEXT"],
+    outputModalities: ["TEXT"],
+    config: o3MiniConfig,
+    systemPromptSupport: true,
+  },
+  {
+    provider: "AI21",
+    id: "ai21.jamba-instruct-v1:0",
+    name: "Jamba-Instruct",
+    inputCostPerToken: 0.0005 / 1e3,
+    outputCostPerToken: 0.0007 / 1e3,
+    inputModalities: ["TEXT"],
+    outputModalities: ["TEXT"],
+    config: ai21JambaInstructConfig,
+    systemPromptSupport: true,
+  },
+  {
+    provider: "AI21",
+    id: "ai21.jamba-1-5-mini-v1:0",
+    name: "Jamba 1.5 Mini",
+    inputCostPerToken: 0.0002 / 1e3,
+    outputCostPerToken: 0.0004 / 1e3,
+    inputModalities: ["TEXT"],
+    outputModalities: ["TEXT"],
+    config: ai21JambaInstructConfig,
+    systemPromptSupport: true,
+  },
+  {
+    provider: "AI21",
+    id: "ai21.jamba-1-5-large-v1:0",
+    name: "Jamba 1.5 Large",
+    inputCostPerToken: 0.002 / 1e3,
+    outputCostPerToken: 0.008 / 1e3,
+    inputModalities: ["TEXT"],
+    outputModalities: ["TEXT"],
+    config: ai21JambaInstructConfig,
+    systemPromptSupport: true,
+  },
+  {
+    provider: "Writer",
+    id: "us.writer.palmyra-x4-v1:0",
+    name: "Palmyra X4",
+    region: "us-west-2",
+    inputModalities: ["TEXT"],
+    outputModalities: ["TEXT"],
+    config: writerModelConfig,
+    systemPromptSupport: true,
+  },
+  {
+    provider: "Writer",
+    id: "us.writer.palmyra-x5-v1:0",
+    name: "Palmyra X5",
+    region: "us-west-2",
+    inputModalities: ["TEXT"],
+    outputModalities: ["TEXT"],
+    config: writerModelConfig,
+    systemPromptSupport: true,
+  },
+];
+
+export const imageModels: ImageModel[] = [
+  {
+    inputModalities: ["TEXT", "IMAGE"],
+    id: "amazon.titan-image-generator-v1",
+    name: "Titan G1 V1",
+    provider: "Amazon",
+    config: titanImageModelConfig,
+  },
+  {
+    inputModalities: ["TEXT", "IMAGE"],
+    id: "amazon.titan-image-generator-v2:0",
+    name: "Titan G1 V2",
+    provider: "Amazon",
+    config: titanImageModelConfig,
+  },
+  {
+    provider: "Amazon",
+    id: "amazon.nova-canvas-v1:0",
+    name: "Nova Canvas",
+    inputModalities: ["TEXT"],
+    config: titanImageModelConfig,
+  },
+  {
+    provider: "Stability AI",
+    id: "stability.stable-image-core-v1:1",
+    name: "Stable Image Core",
+    inputModalities: ["TEXT"],
+    region: "us-west-2",
+    config: stabilityModelConfig,
+  },
+  {
+    provider: "Stability AI",
+    id: "stability.sd3-5-large-v1:0",
+    name: "Stable Diffusion 3.5 Large",
+    inputModalities: ["TEXT"],
+    region: "us-west-2",
+    config: stabilityModelConfig,
+  },
+  {
+    provider: "Stability AI",
+    id: "stability.stable-image-ultra-v1:1",
+    name: "Stable Image Ultra",
+    inputModalities: ["TEXT"],
+    region: "us-west-2",
+    config: stabilityModelConfig,
+  },
+];
+
+export const videoModels: VideoModel[] = [
+  {
+    provider: "Amazon",
+    id: "amazon.nova-reel-v1:0",
+    name: "Nova Reel",
+    costPerSecond: 0.08,
+    config: [],
+  },
+  {
+    provider: "Luma Labs",
+    id: "luma.ray-v2:0",
+    name: "Luma Ray v2",
+    region: "us-west-2",
+    config: lumaRay2ModelConfig,
+  },
+];
